Replace QPointer with QCPPointer
On Qt4, there is is a horrible performance problem when removing a
large number of items from the graph. This happens when closing a
tracefile. I have traced this to the Qt4 implementation of QPointer,
especially its usage of the guards in QMetaObject.

The solution is to copy the QPointer implementation from Qt5 and
call it QCPPointer. This way, the performance problem vanishes
on Qt4.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>