QCPLayer: Use QCPList instead of QList
The mChildren member for QCPLayer becomes a problem when we have a
very large number of plottables. For example the QList<T>::contains()
operation is linear with respect to the number of elements.

In order to improve the situation, with a fairly small effort, we
introduce this new QCPList<T> class. It is a doubly linked list
but it also contains a QMultiMap<> that can be used to implement
faster contains() and removeOne() operations.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>