QCPLayer: Use QCPList instead of QList
The mChildren member for QCPLayer becomes a problem when we have a
very large number of plottables. For example the QList<T>::contains()
operation is linear with respect to the number of elements.

In order to improve the situation, with a fairly small effort, we
introduce this new QCPList<T> class. It is a doubly linked list
but it also contains a QMultiMap<> that can be used to implement
faster contains() and removeOne() operations.

It would be possible to use a QMap<> instead of QMultiMap<> but then the
list would have the limitations that you could not insert the same value
twice. This assumption would as far as I can tell hold true for
QCustomPlot, since the lists are now only used for inserting pointers
to objects that only should appear once. However, since there appears to
be no measureable performance penalty for this slightly unecessary
generalism, I have preferred to use this slightly more general approach in
order to keep as much compatibility with QList<> as possible.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>