Make QCustomPlot::mItems use QCPList
This is an attempt to replace QList with QCPList in the
QCustomPlot::mItems variable. This is mainly done for reasons
of performance.

Observe, that in order to make this change, the following two calls
will get worse performance and have been marked as deprecated:

QCPAbstractItem *QCustomPlot::item(int index) const;
bool QCustomPlot::removeItem(int index);

These are not used by traceshark, not strictly necessary and would be
hard to implement in an efficient way.

Signed-off-by: Viktor Rosendahl <viktor.rosendahl@gmail.com>